Adding BaseEntity and UserStatus along with several services for Users. #06-11

+++ OnlineShop.Application/Services/Users/Commands/Create/RolesInCreateUserRequestInfo.cs
@@ -2,6 +2,6 @@
 {
     public class RolesInCreateUserRequestInfo
     {
-        public byte Id { get; set; }
+        public ulong Id { get; set; }
     }
 }
